+++ NEWS.md
@@ -2,6 +2,11 @@
 
 * fixed classes of returned dataframe columns
 * added documentation of returned dataframe columns
+* added data object `weekdays`, a numeric vector storing numeric equivalents of named weekdays, 
+including start and end days for the CDC Epidemiological week and ISO week.
+* added function `round2wday()` to round dates to nearest weekday (specified as int, named weekday, 
+start or end of CDC epiweek, or start or end of ISO week.)
+* added `fixdates` parameter to `tycho2()`
 
 # tycho2 0.0.1